+++ packages/common-library/src/common_library/groups_enums.py
@@ -1,0 +1,13 @@
+import enum
+
+
+class GroupType(enum.Enum):
+    """
+    standard: standard group, e.g. any group that is not a primary group or special group such as the everyone group
+    primary: primary group, e.g. the primary group is the user own defined group that typically only contain the user (same as in linux)
+    everyone: the only group for all users
+    """
+
+    STANDARD = "standard"
+    PRIMARY = "primary"
+    EVERYONE = "everyone"
