+++ packages/models-library/src/models_library/groups.py
@@ -1,7 +1,7 @@
-import enum
 from typing import Annotated, Final, NamedTuple, TypeAlias
 
 from common_library.basic_types import DEFAULT_FACTORY
+from common_library.groups_enums import GroupType as GroupTypeEnum
 from pydantic import BaseModel, ConfigDict, EmailStr, Field, field_validator
 from pydantic.types import PositiveInt
 from typing_extensions import TypedDict
@@ -14,27 +14,14 @@ EVERYONE_GROUP_ID: Final[int] = 1
 
 GroupID: TypeAlias = PositiveInt
 
-
-__all__: tuple[str, ...] = ("GroupID",)
-
-
-class GroupTypeInModel(str, enum.Enum):
-    """
-    standard: standard group, e.g. any group that is not a primary group or special group such as the everyone group
-    primary: primary group, e.g. the primary group is the user own defined group that typically only contain the user (same as in linux)
-    everyone: the only group for all users
-    """
-
-    STANDARD = "standard"
-    PRIMARY = "primary"
-    EVERYONE = "everyone"
+__all__: tuple[str, ...] = ("GroupTypeEnum",)
 
 
 class Group(BaseModel):
     gid: PositiveInt
     name: str
     description: str
-    group_type: Annotated[GroupTypeInModel, Field(alias="type")]
+    group_type: Annotated[GroupTypeEnum, Field(alias="type")]
     thumbnail: str | None
 
     inclusion_rules: Annotated[
@@ -45,7 +32,7 @@ class Group(BaseModel):
     ] = DEFAULT_FACTORY
 
     _from_equivalent_enums = field_validator("group_type", mode="before")(
-        create_enums_pre_validator(GroupTypeInModel)
+        create_enums_pre_validator(GroupTypeEnum)
     )
 
     model_config = ConfigDict(populate_by_name=True)
